% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbf_crawler.R
\name{cbf_crawler}
\alias{cbf_crawler}
\title{cbf_crawler}
\usage{
cbf_crawler(query, delay_max = 1)
}
\arguments{
\item{query}{'character', "A vector consisting of one or more InChIKeys.}

\item{delay_max}{'number', "crawl delay time, defult = 1, To prevent excessive load on the website, we set the minimum value of the crawl delay to 0.5 seconds, and the maximum value represents a random variation in the crawl delay within the range of 0.5 seconds to that value.}
}
\value{
A dataframe, contains InChIKey, kingdom, superclass, class, subclass parent_levels and description.
}
\description{
Compound classyfication based on cbf of Fiehn lab.
}
\examples{
 
\dontrun{
# demo
xx <- c("RVQCZHZIMZMGAD-IDFNXBDGNA-N","UKMSUNONTOPOIO-UHFFFAOYSA-N","DPUOLQHDNGRHBS-KTKRTIGZSA-N")
library(progressr)
handlers(global = FALSE)
handlers(handler_pbcol(
 adjust = 1.0,
 complete = function(s) cli::bg_red(cli::col_black(s)),
 incomplete = function(s) cli::bg_cyan(cli::col_black(s))
))
# run cbf_crawler
classyfire_tbl = cbf_crawler(query = xx, delay_max = 2)
}
}
\references{
See https://cfb.fiehnlab.ucdavis.edu/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
