% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.classification.R
\name{classfireR2tbl}
\alias{classfireR2tbl}
\title{classfireR2tbl}
\usage{
classfireR2tbl(x)
}
\arguments{
\item{x}{'character' ; result of shawn_get_classification}
}
\value{
pubchem_item An data frame about stored the pubchem information of input cids.
}
\description{
merge classyfireR result
}
\references{
See classyfireR https://github.com/aberHRML/classyfireR
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
