% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.DM_analysis.R
\name{DAM_analysis}
\alias{DAM_analysis}
\title{DAM_analysis}
\usage{
DAM_analysis(
  x,
  left_index,
  right_index,
  left,
  right,
  method = c("t-test", "wilcox-test"),
  method2 = c("opls-da", "pls-da"),
  method3 = p.adjust.methods,
  paired = FALSE
)
}
\arguments{
\item{x}{'Dataframe' or 'matrix'; Peak area file contains all sample.}

\item{left_index}{'characters'; The first sample set names corresponding to the peak area matrix}

\item{right_index}{'characters'; The second sample set names corresponding to the peak area matrix}

\item{left}{'characters'; The group name of first sample}

\item{right}{'characters'; The group name of first sample}

\item{method}{'characters'; t-test， wilcox-test. default is t-test.}

\item{method2}{'characters'; pls-da or opls-da. default is opls-da. By default, 7-fold cross validation were used followed by ropls, if sample size is less than 7, use LOOCV method.}

\item{method3}{'characters'; method for adjust p value, see p.adjust.methods}

\item{paired}{'logical'; paired t-test/wilcox_test or not; default = FASLE}
}
\value{
A list; MDA_tbl is a dataframe contains all compounds and all DAM judge values include log2fc,pvalue,FDR and FDR. opls_out is opls_da result generated by ropls package.
}
\description{
Different accumulation analysis
}
\examples{
\dontrun{
#> For mass_dataset
mat = object \%>\% extract_expression_data() \%>\% rownames_to_column("variable_id")
sample_info = object \%>\% extract_sample_info()
left = "case"
right = "control"
left_index = sample_info \%>\% filter(group == left) \%>\% pull(sample_id)
right_index = sample_info \%>\% filter(group == right) \%>\% pull(sample_id)
res.case_vs_control = DAM_analysis(
  x = mat,
  left = left,
  right = right,
  left_index = left_index,
  right_index = right_index,
  method = "wilcox_test",
  method2 = "opls-da",
  method3 = "fdr",
  paired = FALSE
)

#> for peak picking table 
#> generate a in-silico metabolomics expression matrix
base_expression <- matrix(rnorm(1000, mean = 10, sd = 3), nrow = 1000, ncol = 1)
group1_variability <- matrix(rnorm(3000, mean = 0, sd = 0.5), nrow = 1000, ncol = 3)
group2_variability <- matrix(rnorm(3000, mean = 0, sd = 0.5), nrow = 1000, ncol = 3)
mat <- cbind(base_expression + group1_variability, base_expression + group2_variability)
colnames(mat) = paste0("sample_0",c(1:6))
rownames(mat) = paste0("compound_",c(1:1000))
left_index = c("sample_01","sample_02","sample_03")
right_index = c("sample_04","sample_05","sample_06")
left = "case"
right = "control"

res.case_vs_control = DAM_analysis(
  x = mat,
  left = left,
  right = right,
  left_index = left_index,
  right_index = right_index,
  method = "wilcox_test",
  method2 = "opls-da",
  method3 = "fdr",
  paired = FALSE
)
}
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
