% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.DM_report.R
\name{DAM_report}
\alias{DAM_report}
\title{DM_report}
\usage{
DAM_report(
  working_dir = "./",
  peak_path,
  group_path,
  annotation_path,
  left,
  right,
  pval_cut = 0.05,
  qval_cut = 1,
  VIP_cut = 1,
  log2fc_cut = 0,
  Interactive = TRUE
)
}
\arguments{
\item{working_dir}{'characters'; working dir path.}

\item{peak_path}{'characters'; peak matrix name}

\item{group_path}{'characters'; group file name}

\item{annotation_path}{'characters'; annotation_file}

\item{left}{'characters'; The group name of first sample}

\item{right}{'characters'; The group name of first sample}

\item{pval_cut}{'numeric'; pvalue cutoff from 0-1, default = 0.05}

\item{qval_cut}{'numeric'; qvalue (FDR) cutoff from 0-1, default = 0.05}

\item{VIP_cut}{'numeric'; VIP of opls-da or pls-da cutoff, default = 1}

\item{log2fc_cut}{numeric'; log2(foldchange) cutoff, default = 0}
}
\value{
A list; MDA_tbl is a dataframe contains all compounds and all DAM judge values include log2fc,pvalue,FDR and FDR. opls_out is opls_da result generated by ropls package.
}
\description{
Get kegg cid by CTS
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
