% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAM_volcano_plotly.R
\name{DAM_volcano_plotly}
\alias{DAM_volcano_plotly}
\title{DAM_volcano_plotly}
\usage{
DAM_volcano_plotly(
  x,
  title,
  pval_cut = 0.05,
  log2fc_cut = 1,
  qval_cut = 0.05,
  VIP_cut = 1
)
}
\arguments{
\item{x}{'dataframe' DAM analysis result of MDAtoolkits::DM_analysis.}

\item{title}{"character', title for volcano plot.}

\item{pval_cut}{'numeric'; pvalue cutoff, default = 0.05}

\item{log2fc_cut}{'numeric'; log2fc cutoff, default = 1}

\item{qval_cut}{'numeric'; FDR / adj.p or qvalue cutoff, default = 0.05}

\item{VIP_cut}{'numeric'; VIP cutoff, default = 1}
}
\value{
ggplot2 object
}
\description{
Interactive volcano plot for DAM analysis result.
}
\examples{
 
\dontrun{
# import file
mat = step7_result \%>\% select("CompoundID",starts_with("sample")) \%>\% as.matrix
left_index = c("sample_01","sample_02","sample_03")
right_index = c("sample_04","sample_05","sample_06")
left = "case"
right = "control"
res_dm <- DM_analysis(x = mat,left_index,right_index,left,right)
}
DAM_volcano_plotly(x = res_dm$Diff_result_clean)
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
