% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.DM_analysis.R
\name{DM_analysis}
\alias{DM_analysis}
\title{DM_analysis}
\usage{
DM_analysis(
  x,
  left_index,
  right_index,
  left,
  right,
  method = "t-test",
  method2 = "opls-da"
)
}
\arguments{
\item{x}{'Dataframe' or 'matrix'; Peak area file contains all sample.}

\item{left_index}{'characters'; The first sample set names corresponding to the peak area matrix}

\item{right_index}{'characters'; The second sample set names corresponding to the peak area matrix}

\item{left}{'characters'; The group name of first sample}

\item{right}{'characters'; The group name of first sample}

\item{method}{'characters'; t-test or anova-test. default is t-test.}

\item{method2}{'characters'; pls-da or opls-da. default is opls-da.}
}
\value{
A list; MDA_tbl is a dataframe contains all compounds and all DAM judge values include log2fc,pvalue,FDR and FDR. opls_out is opls_da result generated by ropls package.
}
\description{
Different accumulation analysis
}
\examples{
\dontrun{
# set param
mat = step7_result \%>\% select("CompoundID",starts_with("sample")) \%>\% as.matrix
left_index = c("sample_01","sample_02","sample_03")
right_index = c("sample_04","sample_05","sample_06")
left = "case"
right = "control"
DM_analysis(x = mat,left_index,right_index,left,right)
}
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
