% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_selection.R
\name{extract_fragment}
\alias{extract_fragment}
\title{extract_fragment}
\usage{
extract_fragment(ms2_data, vari_id, spec_id)
}
\arguments{
\item{ms2_data}{'list', ms2 information in mass_dataset object}

\item{vari_id}{'character', variable_id of features have ms2 information}

\item{spec_id}{'character', spectrum id}
}
\value{
A dataframe, with MRM selection result.
}
\description{
Running mrm selection for tidymass workflow.
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
