% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.CTS_kegg.R
\name{InChikey2kegg}
\alias{InChikey2kegg}
\title{InChikey2kegg}
\usage{
InChikey2kegg(url)
}
\arguments{
\item{query}{'character' or 'data.frame'; For single InChiKey, query equals the InChiKey, For multiple InChiKeys, query is a data frame from mda_pubchem_crawler.}
}
\value{
KEGGID InChikey 2 kegg cid.
}
\description{
Get kegg cid by CTS
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
