% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.classification.R
\name{mda_classfire}
\alias{mda_classfire}
\title{mda_classfire.}
\usage{
mda_classfire(query, type = "multiple")
}
\arguments{
\item{query}{'character' or 'data.frame'; For single InChiKey, query equals the InChiKey, For multiple InChiKeys, query is a data frame from mda_pubchem_crawler.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}
}
\value{
pubchem_item An data frame about stored the pubchem information of input cids.
}
\description{
Compound classification via classyfireR
}
\references{
See classyfireR https://github.com/aberHRML/classyfireR
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
