% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.CTS_kegg.R
\name{mda_CTS_kegg}
\alias{mda_CTS_kegg}
\title{mda_CTS_kegg.}
\usage{
mda_CTS_kegg(query, type = "multiple", key = "InChIKey", core_num = 1)
}
\arguments{
\item{query}{'character' or 'data.frame'; For single InChiKey, query equals the InChiKey, For multiple InChiKeys, query is a single column of input, InChIKey or name.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}

\item{key}{'character'; InChIKey or name?}

\item{core_num}{'numeric'; how many cores you want to use}
}
\value{
KEGGID InChikey 2 kegg cid.
}
\description{
Get kegg cid by CTS
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
