% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.mda_get_cid.R
\name{mda_data_org}
\alias{mda_data_org}
\title{mda_data_org.}
\usage{
mda_data_org(compound_info, source = "CD")
}
\arguments{
\item{compound_info}{A compound table input. contains at least five columns. 1. Customized Compound_ID, 2. compound name, 3. mz, 4. mf, 5.RT, for CD result add Judge.}

\item{source}{Input compound information from Compound discovery or other software.}
}
\value{
cid_tbl An organized data for following steps.
}
\description{
Data organization of input data.
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
}
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
