% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.mda_get_cid.R
\name{mda_get_cid}
\alias{mda_get_cid}
\title{mda_get_cid}
\usage{
mda_get_cid(
  data_info,
  type = "multiple",
  match = "all",
  core_num = 8,
  from = "name"
)
}
\arguments{
\item{data_info}{A vector or data.frame; For single compound, just set the compound name as data_info, for multipule compounds, use organized data by mda_data_org.}

\item{type}{'multiple','single', For one compound or multiple compounds.}

\item{match}{'all','first'; Check webchem::get_cid match, return all matched cid or the first match, default 'all' .}

\item{core_num}{'numeric'; how many cores you want to use}

\item{from}{'character'; see webchem::get_cid}
}
\value{
cid_tbl A table which webchem::get_cid generated.
}
\description{
Get cid by compound name, single or multiple.
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mad_get_cid(data_info = orz_data)
# please wait for a while
}
}
\references{
See webchem https://github.com/ropensci/webchem
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
