% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_get_cid_fast.R
\name{mda_get_cid_fast}
\alias{mda_get_cid_fast}
\title{mda_get_cid_fast.}
\usage{
mda_get_cid_fast(query, probe = "name", core_num = 1)
}
\arguments{
\item{query}{A single compound name or multiple compounds}

\item{probe}{'character', form of query, name or inchikey.}

\item{core_num}{'numeric'; how many cores you want to use}
}
\value{
cid_tbl contains cid and InChIKey.
}
\description{
Data organization of input data.
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mad_get_cid(query = orz_data)
# please wait for a while
}

}
\references{
See webchem https://github.com/ropensci/webchem
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
