% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_knapsack_spider.R
\name{mda_knapsack_spider}
\alias{mda_knapsack_spider}
\title{mda_knapsack_spider}
\usage{
mda_knapsack_spider(input_data, type = "multiple", core_num = 8)
}
\arguments{
\item{input_data}{A vector or data.frame; a data.frame with one of the column named knapsack.}

\item{type}{'multiple','single', For one compound or multiple compounds.}

\item{core_num}{'numeric'; how many cores you want to use}
}
\value{
A dataframe, contains C_ID, C_name, formula, mw, InChIKey SMILE, synonyms and organism.
}
\description{
Get compound information by CID of KNApSAcK database.
}
\examples{
 
\dontrun{
# import file
xx <- data.frame(knapsack=c(""))
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mad_get_cid(data_info = orz_data)
# please wait for a while
}
}
\references{
See webchem https://github.com/ropensci/webchem
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
