% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.classification.R
\name{mda_merge_info}
\alias{mda_merge_info}
\title{mda_merge_info}
\usage{
mda_merge_info(orz_data, name2cid, pubchem_detail, pubchem_class, CTS_kegg)
}
\arguments{
\item{orz_data}{'dataframe'; generated by function mda_data_org.}

\item{name2cid}{'dataframe'; generated by function mda_get_cid.}

\item{pubchem_detail}{'dataframe'; generated by function mda_pubchem_crawler.}

\item{pubchem_class}{'dataframe'; generated by function mda_classfire.}
}
\value{
pubchem_item An data frame about stored the pubchem information of input cids.
}
\description{
Merge all information.
}
\examples{
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mda_get_cid(data_info = orz_data)
# please wait for a while
pubchem_detail = mda_pubchem_crawler(cid_info = name2cid,type = "multiple",multi_core = T)
pubchem_class = mda_classfire(query = pubchem_detail,type = "multiple")
final_tbl = mda_merge_info(orz_data,name2cid,pubchem_detail,pubchem_class)
}
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email {shawnwang2016@126.com}.
}
