% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_name2kegg.R
\name{mda_name2kegg}
\alias{mda_name2kegg}
\title{MDA_name2kegg}
\usage{
mda_name2kegg(query, type = "multiple", core_num = 8)
}
\arguments{
\item{query}{'character' or 'data.frame'; For single InChiKey, query equals the InChiKey, For multiple InChiKeys, query is a single column of input, InChIKey or name.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}

\item{core_num}{'numeric'; how many cores you want to use}
}
\value{
KEGGID compound name 2 kegg cid.
}
\description{
Get kegg cid by kegg website
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
