% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.mda_pca.R
\name{mda_pca}
\alias{mda_pca}
\title{mda_PCA}
\usage{
mda_pca(
  mat,
  group,
  interactive = F,
  removeVar = 0.1,
  logTransfer = F,
  title = "test"
)
}
\arguments{
\item{mat}{'Dataframe' or 'matrix'; with Compound_ID as 1st column,.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}
}
\value{
KEGGID InChikey 2 kegg cid.
}
\description{
Get kegg cid by CTS
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
