% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.pubchem_crawler.R
\name{mda_pubchem_crawler}
\alias{mda_pubchem_crawler}
\title{mda_pubchem_crawler.}
\usage{
mda_pubchem_crawler(cid_info, type = "multiple", multi_core = TRUE)
}
\arguments{
\item{cid_info}{'character' or 'data.frame'; For single cid, cid_info equals the pubchem CID, For multiple cid, cid_info is a dataframe generated from mda_get_cid.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}

\item{multi_core}{'TRUE' or 'FALSE'; Run web crawler with multicore or not.}
}
\value{
pubchem_item An data frame about stored the pubchem information of input cids.
}
\description{
Get compound information according the cid by webchem.
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mda_get_cid(data_info = orz_data)
# please wait for a while
pubchem_detail = mda_pubchem_crawler(cid_info = name2cid,type = "multiple",multi_core = T)
}
}
\references{
See webchem https://github.com/ropensci/webchem.
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
