% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.pubchem_crawler.R
\name{mda_pubchem_crawler_sh}
\alias{mda_pubchem_crawler_sh}
\title{mda_pubchem_crawler_sh.}
\usage{
mda_pubchem_crawler_sh(cid)
}
\arguments{
\item{cid}{'character'; pubchem cid}
}
\value{
pubchem_item An data frame about stored the pubchem information of input cids.
}
\description{
Get compound information according the cid by webchem.
}
\references{
See webchem https://github.com/ropensci/webchem.
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
