% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.classification.R
\name{mda_tbl_merge}
\alias{mda_tbl_merge}
\title{mda_tbl_merge}
\usage{
mda_tbl_merge(tbl1, tbl2, col1, col2, method)
}
\arguments{
\item{tbl1}{'dataframe'; table1.}

\item{tbl2}{'dataframe'; table2.}

\item{col1}{'character'; Key column name for table1.}

\item{col2}{'character'; Key column name for table2.}

\item{method}{'character'; left_join, right_join, inner_join or full_join.}
}
\description{
Merge two tables for shinyapp
}
\examples{
 
\dontrun{
newtbl = mda_tbl_merge(tbl1 = orz_data,tbl2 = name2cid,col1 = "query",col2 = "name")
}
}
\references{
https://github.com/gadenbuie/tidyexplain#relational-data
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email {shawnwang2016@126.com}.
}
