% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_selection.R
\name{mrm_selection_cd}
\alias{mrm_selection_cd}
\title{mrm_selection}
\usage{
mrm_selection_cd(x)
}
\arguments{
\item{x}{'dataframe', "A dataframe exported by CD software}
}
\value{
A dataframe, with MRM selection result.
}
\description{
Running mrm selection for compound discoverer result.
}
\examples{
 
\dontrun{
# demo
x = readxl::read_xlsx("~/GeekCloud/TTQC.xlsx",sheet = 2)
out <- mrm_selection_cd(x = x)
}
}
\references{
See
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
