% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_selection.R
\name{oneStepMRMselection}
\alias{oneStepMRMselection}
\title{mrm_selection_tidymass}
\usage{
oneStepMRMselection(obj_ms2)
}
\arguments{
\item{obj_ms2}{'mass_dataset', a mass_dataset object with ms2 annotation}
}
\value{
A dataframe, with MRM selection result.
}
\description{
Running mrm selection for tidymass workflow.
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
