% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.mda_get_cid.R
\name{shawn_get_cid_sh}
\alias{shawn_get_cid_sh}
\title{shawn_get_cid_sh}
\usage{
shawn_get_cid_sh(name, mod)
}
\arguments{
\item{name}{compound name.}

\item{mod}{match model; see webchem .}
}
\value{
cid_tbl A table which webchem::get_cid generated.
}
\description{
get cid for shinyapp
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mad_get_cid(data_info = orz_data)
# please wait for a while
}
}
\references{
See webchem https://github.com/ropensci/webchem
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
