% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.CTS_kegg.R
\name{mda_CTS_kegg}
\alias{mda_CTS_kegg}
\title{mda_CTS_kegg.}
\usage{
mda_CTS_kegg(query, type = "multiple")
}
\arguments{
\item{query}{'character' or 'data.frame'; For single InChiKey, query equals the InChiKey, For multiple InChiKeys, query is a data frame from mda_pubchem_crawler.}

\item{type}{"single','multiple';  For one compound or multiple compounds.}
}
\value{
KEGGID InChikey 2 kegg cid.
}
\description{
Get kegg cid by CTS
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
