% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_get_cid_fast.R
\name{mda_get_cid_fast}
\alias{mda_get_cid_fast}
\title{mda_get_cid_fast.}
\usage{
mda_get_cid_fast(data_info, type = "multiple", core_num = 1)
}
\arguments{
\item{data_info}{A compound table input. Must have atleast two columns, 1st Compound_ID, 2nd CAS.ID or compound name.}

\item{type}{'multiple','single', For one compound or multiple compounds.}

\item{core_num}{'numeric'; how many cores you want to use}
}
\value{
cid_tbl contains cid and InChIKey.
}
\description{
Data organization of input data.
}
\examples{
 
\dontrun{
# import file
xx <- read.csv("cd_result.csv")
# data orgnazation
orz_data = mda_data_org(compound_info = xx, source = "CD")
# Get cid (batch)
name2cid = mad_get_cid(data_info = orz_data)
# please wait for a while
}

}
\references{
See webchem https://github.com/ropensci/webchem
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>.
\email{shawnwang2016@126.com}
}
